공지추가,수정화면 CSS수정&JS추가,환자VO 몸무게 데이터타입(float)변경&보호자 탈퇴여부 데이터타입(boolean)변경

+++ AnimalHospitalProject/src/main/java/kr/human/anihospital/vo/AnimalVO.java
@@ -16,7 +16,7 @@ public class AnimalVO {
 	private String animalName;
 	private String animalType;
 	private String animalSize;
-	private int animalWeight;
+	private float animalWeight;
 	private Date animalAge;
 	private boolean animalGender;
 	private String animalImportantSymptom;
